Revert "Release 20251110 aos post layout rebuild image captions"
