+++ package.json
@@ -18,12 +18,5 @@
     "npm-run-all": "^4.1.5",
     "postcss": "^8.5.6",
     "tailwindcss": "^3.4.17"
-  },
-  "dependencies": {
-    "aos": "^2.3.4",
-    "markdown-it": "^14.1.0",
-    "markdown-it-anchor": "^9.2.0",
-    "markdown-it-attrs": "^4.3.1",
-    "markdown-it-image-figures": "^2.1.1"
   }
 }
