--- src/_includes/partials/accent-bar.njk
@@ -1,4 +1,0 @@
-<div class="relative h-1 w-20 ml-[-0.25rem] md:w-28 md:ml-[-0.35rem]">
-    <div class="absolute inset-0 brand-gradient-right rounded-full mx-auto w-[90%]"></div>
-    <div class="absolute inset-0 brand-gradient-right rounded-full blur animate-[pulse_2s_cubic-bezier(0.4,_0,_0.6,_1)_infinite;]"></div>
-</div>
