+++ src/assets/css/typescale.css
@@ -5,39 +5,39 @@
   --font-family: "Inter", system-ui, -apple-system, Segoe UI, Roboto, "Helvetica Neue", Arial, "Noto Sans", "Liberation Sans", sans-serif;
 
   /* Type Scale */
-  --display-lg-size: 50px;
+  --display-lg-size: 50.52px;
   --display-lg-line-height: 1.5;
   --display-lg-letter-spacing: 0.015em;
 
-  --display-md-size: 44px;
+  --display-md-size: 43.75px;
   --display-md-line-height: 1.55;
   --display-md-letter-spacing: 0.03em;
 
-  --h1-size: 38px;
+  --h1-size: 37.9px;
   --h1-line-height: 1.55;
   --h1-letter-spacing: 0.03em;
 
-  --h2-size: 29px;
+  --h2-size: 28.43px;
   --h2-line-height: 1.55;
   --h2-letter-spacing: 0.03em;
 
-  --h3-size: 26px;
+  --h3-size: 24.62px;
   --h3-line-height: 1.55;
   --h3-letter-spacing: 0.03em;
 
-  --h4-size: 22px;
+  --h4-size: 21.33px;
   --h4-line-height: 1.6;
   --h4-letter-spacing: 0.038em;
 
-  --h5-size: 19px;
+  --h5-size: 19.85px;
   --h5-line-height: 1.6;
   --h5-letter-spacing: 0.045em;
 
-  --h6-size: 18px;
+  --h6-size: 18.47px;
   --h6-line-height: 1.6;
   --h6-letter-spacing: normal;
 
-  --body-xl-size: 17px;
+  --body-xl-size: 17.19px;
   --body-xl-weight: 400;
   --body-xl-line-height: 1.6;
   --body-xl-letter-spacing: normal;
@@ -47,17 +47,17 @@
   --body-line-height: 1.6;
   --body-letter-spacing: normal;
 
-  --body-sm-size: 15px;
+  --body-sm-size: 14.89px;
   --body-sm-weight: 400;
   --body-sm-line-height: 1.65;
   --body-sm-letter-spacing: normal;
 
-  --caption-lg-size: 14px;
+  --caption-lg-size: 13.86px;
   --caption-lg-weight: 400;
   --caption-lg-line-height: 1.65;
   --caption-lg-letter-spacing: normal;
 
-  --caption-size: 13px;
+  --caption-size: 12.9px;
   --caption-weight: 400;
   --caption-line-height: 1.65;
   --caption-letter-spacing: normal;
@@ -67,7 +67,7 @@
   --caption-sm-line-height: 1.65;
   --caption-sm-letter-spacing: normal;
 
-  --overline-size: 11px;
+  --overline-size: 11.17px;
   --overline-weight: 500;
   --overline-line-height: 1.65;
   --overline-letter-spacing: normal;
